requirements.txt
amqp==5.2.0
asgiref==3.7.2
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
billiard==4.2.0
boto3==1.34.28
botocore==1.34.28
celery==5.3.6
certifi==2024.2.2
cffi==1.16.0
channels==4.0.0
channels-redis==4.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
constantly==23.10.4
construct==2.5.3
contourpy==1.2.0
cron-descriptor==1.4.3
cryptography==41.0.7
cycler==0.12.1
Cython==3.0.8
daphne==4.0.0
Django==4.2.10
django-celery-beat==2.6.0
django-celery-results==2.5.1
django-filter==23.5
django-haystack==3.2.1
django-ratelimit==4.1.0
django-storages==1.14.2
django-stubs-ext==4.2.7
django-timezone-field==6.1.0
django-user-agents==0.4.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.1
fonttools==4.49.0
ftfy==6.1.3
h5py==3.10.0
haystack==0.42
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
inflection==0.5.1
jmespath==1.0.1
joblib==1.3.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
kombu==5.3.5
msgpack==1.0.7
packaging==23.2
patsy==0.5.6
pefile==2023.2.7
pillow==10.2.0
prompt-toolkit==3.0.43
psutil==5.9.8
psycopg2-binary==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
pymc3==3.4.1
pyOpenSSL==24.0.0
pyotp==2.9.0
pyparsing==3.1.1
pyrocko==2024.1.10
python-bidi==0.4.2
python-crontab==3.0.0
python-dateutil==2.8.2
python-ptrace==0.9.8
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
referencing==0.32.1
requests==2.31.0
rest-framework-simplejwt==0.0.2
rpds-py==0.17.1
s3transfer==0.10.0
service-identity==24.1.0
setuptools-scm==8.0.4
six==1.16.0
sqlparse==0.4.4
tqdm==4.66.2
Twisted==23.10.0
txaio==23.1.1
types-pytz==2024.1.0.20240203
types-PyYAML==6.0.12.12
typing_extensions==4.9.0
tzdata==2024.1
ua-parser==0.18.0
uritemplate==4.1.1
urllib3==1.26.18
user-agents==2.2.0
vine==5.1.0
wcwidth==0.2.13
Whoosh==2.7.4
zope.interface==6.1
kavenegar==1.1.2
django-cors-headers==4.3.1
django-debug-toolbar